finished on improving the quantity and delete function

+++ app/page.js
@@ -3,7 +3,7 @@
 import { useEffect, useState } from "react";
 import { Box, Stack, Typography, Button, Modal, TextField } from "@mui/material";
 import { firestore } from "@/firebase";
-import { collection, getDocs, query, doc, setDoc, deleteDoc } from "@firebase/firestore";
+import { collection, getDocs, query, doc, setDoc, deleteDoc, getDoc } from "@firebase/firestore";
 
 const style = {
   position: 'absolute',
@@ -34,7 +34,7 @@ export default function Home() {
       const docs = await getDocs(snapshot)
       const pantryTrackerList = []
       docs.forEach((doc) => {
-        pantryTrackerList.push(doc.id)
+        pantryTrackerList.push({name: doc.id, ...doc.data()})
       })
       console.log(pantryTrackerList)
       setPantryTracker(pantryTrackerList)
@@ -47,15 +47,30 @@ export default function Home() {
 
   const addItem = async (item) => {
     const docRef = doc(collection(firestore, 'pantryTracker'), item)
-    await setDoc(docRef, {})
-    await updatePantryTracker()
+    const docSnap = await getDoc(docRef)
+    if (docSnap.exists()) {
+      const {count} = docSnap.data()
+      await setDoc(docRef, {count: count + 1})
+      
+    } else{
+      await setDoc(docRef, {count:1})
+      
+    }  
+    await updatePantryTracker() 
   }
 
   const removeItem = async (item) => {
     const docRef = doc(collection(firestore, 'pantryTracker'), item)
-    await deleteDoc(docRef)
+    const docSnap = await getDoc(docRef)
+    if (docSnap.exists()) {
+      const {count} = docSnap.data()
+      if (count == 1) {
+        await deleteDoc(docRef)
+      } else {
+        await setDoc(docRef, {count: count -1})
+      }
+    }
     await updatePantryTracker()
-    
   }
 
   return (
@@ -120,9 +135,9 @@ export default function Home() {
         height="300px" 
         spacing={2} 
         overflow={"auto"}>
-          {pantryTracker.map((i) => (
+          {pantryTracker.map(({name, count}) => (
             <Box
-              key={i}
+              key={name}
               width="100%"
               minHeight="150px"
               display={"flex"}
@@ -137,12 +152,15 @@ export default function Home() {
               textAlign={"center"}>
                 {
                   // capitalize first letter of the item
-                  i.charAt(0).toUpperCase() + i.slice(1)
+                  name.charAt(0).toUpperCase() + name.slice(1)
                 }
+              </Typography>
+              <Typography variant={'h3'} color={'#333'} textAlign={'center'}>
+                Quantity: {count}
               </Typography>
               <Button 
                 variant="contained" 
-                onClick={() =>removeItem(i)}>
+                onClick={() =>removeItem(name)}>
                 Remove
               </Button>
             </Box>
